# docs/swagger.yaml
basePath: /
definitions:
  entity.ExpenseRequest:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      value:
        type: number
    type: object
  entity.ExpenseResponse:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      userId:
        type: string
      value:
        type: number
    type: object
  entity.ReceiptRequest:
    properties:
      date:
        type: string
      description:
        type: string
      value:
        type: number
    type: object
  entity.ReceiptResponse:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      userId:
        type: string
      value:
        type: number
    type: object
  entity.UserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  model.BalanceSumaryResponse:
    properties:
      categoryBalance:
        additionalProperties:
          type: number
        type: object
      monthBalance:
        type: number
      totalExpense:
        type: number
      totalReceipt:
        type: number
    type: object
host: alurachallengebackend2ndedition-env.eba-cmaxmrtx.us-east-2.elasticbeanstalk.com
info:
  contact: {}
  title: Alura Backend Challenge 2nd Edition API
  version: 1.1.1
paths:
  /budget-control/api/v1/balance/{year}/{month}:
    get:
      description: get month balance sumary
      parameters:
      - description: Year
        in: path
        name: year
        required: true
        type: string
      - description: Month
        in: path
        name: month
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BalanceSumaryResponse'
      tags:
      - Sumary
  /budget-control/api/v1/expense:
    get:
      description: Find all expenses
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ExpenseResponse'
            type: array
      summary: Find all expenses
      tags:
      - Expenses
    post:
      description: 'Create a new expense. Obs.: you cannot create two expenses with
        the same description in a single month.'
      parameters:
      - description: Expense
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/entity.ExpenseRequest'
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Create a new expense
      tags:
      - Expenses
  /budget-control/api/v1/expense/{id}:
    delete:
      description: Delete an expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
      summary: Delete an expense
      tags:
      - Expenses
    get:
      description: Find expense by id
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ExpenseResponse'
      summary: Find expense by id
      tags:
      - Expenses
    put:
      description: Update an expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      - description: Expense
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/entity.ExpenseRequest'
      responses:
        "204":
          description: ""
      summary: Update an expense
      tags:
      - Expenses
  /budget-control/api/v1/expense/{year}/{month}:
    get:
      description: Find all expenses by period
      parameters:
      - description: Year
        in: path
        name: year
        required: true
        type: integer
      - description: Month
        in: path
        name: month
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ExpenseResponse'
            type: array
      summary: Find all expenses by period
      tags:
      - Expenses
  /budget-control/api/v1/health:
    get:
      description: return server status
      responses:
        "200":
          description: ""
        "404":
          description: ""
      tags:
      - Health
  /budget-control/api/v1/receipt:
    get:
      description: Find all receipts
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ReceiptResponse'
            type: array
      summary: Find all receipts
      tags:
      - Receipts
    post:
      description: 'Create a new receipt. Obs.: you cannot create two receipts with
        the same description in a single month.'
      parameters:
      - description: Receipt
        in: body
        name: receipt
        required: true
        schema:
          $ref: '#/definitions/entity.ReceiptRequest'
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Create a new receipt
      tags:
      - Receipts
  /budget-control/api/v1/receipt/{id}:
    delete:
      description: Delete a receipt
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
      summary: Delete a receipt
      tags:
      - Receipts
    get:
      description: Find a receipt by id
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ReceiptResponse'
      summary: Find a receipt by id
      tags:
      - Receipts
    put:
      description: Update a receipt
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: string
      - description: Receipt
        in: body
        name: receipt
        required: true
        schema:
          $ref: '#/definitions/entity.ReceiptRequest'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update a receipt
      tags:
      - Receipts
  /budget-control/api/v1/receipt/period/{year}/{month}:
    get:
      description: Find all receipts by Period
      parameters:
      - description: Year
        in: path
        name: year
        required: true
        type: integer
      - description: Month
        in: path
        name: month
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ReceiptResponse'
            type: array
      summary: Find all receipts by Period
      tags:
      - Receipts
  /budget-control/api/v1/user:
    post:
      description: create a new user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.UserRequest'
      tags:
      - User
  /budget-control/api/v1/user/authenticate:
    post:
      description: authenticate user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserRequest'
      responses:
        "201":
          description: token
          schema:
            type: string
      tags:
      - User
swagger: "2.0"